app/requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.2
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dash==1.19.0
dash-bootstrap-components==0.11.1
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
JayDeBeApi==1.2.3
jedi==0.18.0
Jinja2==2.11.2
JPype1==1.2.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
notebook==6.2.0
numpy==1.19.5
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==21.0.1
requests==2.25.1
retrying==1.3.3
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
visdcc==0.0.40
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
./nativeAPI_wheel/irisnative-1.0.0-cp34-abi3-linux_x86_64.whl